pyproject.toml
[project]
name = "my-django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
    "django-allauth>=65.3.1",
    "django-vite>=3.0.6",
    "djangorestframework>=3.15.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 100
exclude = [
    "migrations",
    "node_modules",
    "static",
    "templates",
    "venv",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"